Fix rendering of features listview when it's visible
The alignment: start trick from the parent commit doesn't quite work when the ListView
is shown, as then it still won't take up the remaining space
(not even with vertical-stretch: 1).

One workaround is to unconditionally show it, but that doesn't look
great IMO. So instead a spacing rectangle does the trick.